+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/showdetail/ShowDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.nexters.boolti.presentation.screen.show
+package com.nexters.boolti.presentation.screen.showdetail
 
 import android.content.Intent
 import android.net.Uri
