+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/showdetail/TicketReservationPeriod.kt
@@ -1,4 +1,4 @@
-package com.nexters.boolti.presentation.screen.show
+package com.nexters.boolti.presentation.screen.showdetail
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
